asgi_uk_medical_bot-main/requirements.txt
drf-yasg==1.21.10
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
uvicorn[standard]
celery
redis
django-celery-beat
flower
psycopg2-binary
django-cors-headers>=4.0.0
requests
channels
channels-redis
numpy
pillow
pandas
openpyxl
drf-spectacular
drf-spectacular[sidecar]
scipy==1.14.1
whitenoise
starlette==0.27.0